# BookStore/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/displayBooks": {
            "get": {
                "summary": "Display Books",
                "description": "display ALL Books ",
                "operationId": "3d7cea23db0e68ab3435a184f6478c11",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/displayBookbyID/{id}": {
            "get": {
                "summary": "Display Books by ID",
                "description": "Display Book Based on ID",
                "operationId": "2e2e54c14098df4bd4738fc635182091",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "success"
                    },
                    "401": {
                        "description": "No Book Found with That ID to Display"
                    }
                }
            }
        },
        "/api/deleteBookID/{id}": {
            "delete": {
                "summary": "Delete Book",
                "description": "Delete users Books by ID",
                "operationId": "05655555edea94358171cf7233579f67",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "Book Deleted Successfully"
                    },
                    "401": {
                        "description": "No Book Found with that ID to Delete"
                    }
                }
            }
        },
        "/api/updateBookQuantitybyID": {
            "post": {
                "summary": "Add Quantity to Existing Books",
                "description": "Adding Quantity of Books based on ID",
                "operationId": "7c2aadcf746730c6c7f08ec8f3af826b",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "Description"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quantity of Books Updated Successfully"
                    },
                    "401": {
                        "description": "No Book Found with that ID to Update"
                    }
                }
            }
        },
        "/api/searchBook": {
            "post": {
                "summary": "Search Books",
                "description": "Search Books based on name,id and author",
                "operationId": "3102fe2b9869f32fc1d498fc0e4f9994",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "No book Found with the entered Value"
                    }
                }
            }
        },
        "/api/sortByPriceLowToHigh": {
            "get": {
                "summary": "Sort Books by price Low to High",
                "description": "Sort Books Based on Price from Low to High",
                "operationId": "874fed62ea226895914039fd43b5b231",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "success"
                    },
                    "401": {
                        "description": "No Book Found with That ID to Display"
                    }
                }
            }
        },
        "/api/sortByPriceHighToLow": {
            "get": {
                "summary": "Sort Books by price High to Low",
                "description": "Sort Books Based on Price from High to Low",
                "operationId": "6a7db5f114710572ea8694791e2c470a",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "success"
                    },
                    "401": {
                        "description": "No Book Found with That ID to Display"
                    }
                }
            }
        },
        "/api/changePassword": {
            "post": {
                "summary": "Change Password",
                "description": "Changing Password in Postman",
                "operationId": "f6f18402e7c882d2c41ae5656c9e6c33",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "newPassword"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "newPassword": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "password updated successfully"
                    },
                    "400": {
                        "description": "Check your old password"
                    }
                }
            }
        },
        "/api/forgotPassword": {
            "post": {
                "summary": "forgotPassword ",
                "description": "Resetting Password using forgotPassword function",
                "operationId": "e671da130e82b1c6afe6438a521a38a0",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token Sent to Mail to Reset Password"
                    },
                    "404": {
                        "description": "Email does not exists"
                    }
                }
            }
        },
        "/api/resetPassword": {
            "post": {
                "summary": "Resetting Password",
                "description": "Resetting Password through Token",
                "operationId": "07c8921dbaa412466d32ddd23a56609e",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "token"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "token": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "password Reset successfull"
                    },
                    "401": {
                        "description": "You have entered invalid token"
                    }
                }
            }
        },
        "/api/registration": {
            "post": {
                "summary": "User Registration",
                "description": "To Register the User has to enter his details",
                "operationId": "cedfeeaaba352fe588d7b0a4f7af52ef",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "role",
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "confirm_password"
                                ],
                                "properties": {
                                    "role": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "confirm_password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data Registered succesfully"
                    },
                    "401": {
                        "description": "The Email has already been taken"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "User login",
                "description": "login to UI using Mail and Password",
                "operationId": "0fc9610184083ac6dddab2c63139225a",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login success"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        }
    }
}